requirements.txt
crewai[tools]>=0.95.0,<1.0.0
google-genai>=0.1.0
pillow>=10.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
fastapi
uvicorn
python-multipart
pydantic
setuptools


